lenskart/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.2.0
requests>=2.31.0
pydantic>=2.5.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
httpx>=0.25.2
pandas>=2.2.0
openpyxl>=3.1.2
huggingface_hub>=0.20.0
torch
transformers

